requirements.txt
charset-normalizer==3.0.1
django>=3.2.4,<3.3
django-environ==0.9.0
djangorestframework>=3.12.4,<3.13
requests==2.28.2 
sqlparse==0.4.3 
urllib3==1.26.14
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
gunicorn==20.1.0